FastAPI-backend/requirements.txt
# --- Core Framework ---
fastapi[standard]
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

# --- Database ---
SQLAlchemy
psycopg2-binary
alembic

# --- Auth & Security ---
# passlib[bcrypt]
# python-jose
# bcrypt
cryptography

# --- LLM / AI / Vector Search ---
langchain
langchain-community
sentence-transformers
tiktoken
qdrant-client

# --- File / Text Handling (if chatbot uses documents) ---
PyPDF2
python-docx
beautifulsoup4

# --- Networking & Utils ---
httpx
requests
tenacity
aiofiles
orjson
typing-extensions

# --- Logging ---
loguru

# --- Dev Tools (local only, not production) ---
black
isort
flake8
mypy
pre-commit
